requirements.txt
kafka-python

minio